chore: Update package dependencies
This commit updates the package.json file to include the latest versions of the dependencies. Specifically, it updates "@emotion/react" and "@emotion/styled" to version 11.10.6, "@mui/icons-material" to version 5.11.16, "@mui/lab" to version 5.0.0-alpha.131, "@mui/material" to version 5.12.1, "axios" to version 1.3.6, "dotenv" to version 16.1.3, "eslint-config-react-app" to version 7.0.1, "react" and "react-dom" to version 18.2.0, "react-countup" to version 6.4.2, "react-redux" to version 8.0.5, "react-router-dom" to version 6.10.0, "react-scripts" to version 5.0.1, "recharts" to version 2.6.2, "redux" to version 4.2.1, and "styled-components" to version 5.3.10. This ensures that the project is using the latest stable versions of the dependencies.