feat: Add CustomPieChart component
This commit adds a new component called CustomPieChart. The component is responsible for rendering a pie chart using the recharts library. It accepts a data prop which contains an array of objects representing the data to be displayed in the chart. The component also includes a renderCustomizedLabel function for customizing the labels displayed on the chart. The colors used in the chart are defined in the COLORS constant.